+++ lib/roo_on_rails.rb
@@ -11,5 +11,5 @@ if defined?(Rails)
   require 'roo_on_rails/railties/http'
   require 'roo_on_rails/railties/sidekiq'
   require 'roo_on_rails/railties/rake_tasks'
-  require 'roo_on_rails/railties/google_auth'
+  require 'roo_on_rails/railties/google_oauth'
 end