+++ lib/roo_on_rails/checks/environment.rb
@@ -3,10 +3,7 @@ require 'roo_on_rails/checks/github/branch_protection'
 require 'roo_on_rails/checks/heroku/app_exists'
 require 'roo_on_rails/checks/heroku/preboot_enabled'
 require 'roo_on_rails/checks/heroku/app_exists'
-require 'roo_on_rails/checks/sidekiq/settings'
 require 'roo_on_rails/checks/heroku/drains_metrics'
-require 'roo_on_rails/checks/documentation/playbook'
-require 'roo_on_rails/checks/google_oauth/initializer'
 require 'roo_on_rails/checks/papertrail/all'
 
 module RooOnRails
@@ -15,9 +12,6 @@ module RooOnRails
       requires GitHub::BranchProtection
       requires Heroku::DrainsMetrics
       requires Heroku::PrebootEnabled
-      requires Sidekiq::Settings
-      requires Documentation::Playbook
-      requires GoogleOauth::Initializer
       requires Papertrail::All
 
       def call