+++ lib/roo_on_rails/harness.rb
@@ -2,21 +2,29 @@ require 'thor'
 require 'hashie'
 require 'roo_on_rails/checks/environment'
 require 'roo_on_rails/environment'
+require 'roo_on_rails/checks/sidekiq/settings'
+require 'roo_on_rails/checks/documentation/playbook'
 
 module RooOnRails
   class Harness
     include Thor::Shell
 
-    def initialize(try_fix: false, context: nil)
+    def initialize(try_fix: false, context: Hashie::Mash.new, dry_run: false)
       @try_fix = try_fix
-      @context = context || Hashie::Mash.new
+      @context = context
+      @dry_run = dry_run
     end
 
     def run
-      ENV.fetch('ROO_ON_RAILS_ENVIRONMENTS', 'staging,production').split(',').each do |env|
-        Checks::Environment.new(env: env, fix: @try_fix, context: @context)
+      checks = [
+        Checks::Sidekiq::Settings.new(fix: @try_fix, context: @context, dry_run: @dry_run),
+        Checks::Documentation::Playbook.new(fix: @try_fix, context: @context, dry_run: @dry_run),
+      ]
+      environments.each do |env|
+        checks << Checks::Environment.new(env: env.strip, fix: @try_fix, context: @context, dry_run: @dry_run)
       end
 
+      checks.each(&:run)
       self
     rescue Shell::CommandFailed
       say 'A command failed to run, aborting', %i[bold red]
@@ -24,6 +32,12 @@ module RooOnRails
     rescue Checks::Failure
       say 'A check failed, exiting', %i[bold red]
       exit 1
+    end
+
+    private
+
+    def environments
+      ENV.fetch('ROO_ON_RAILS_ENVIRONMENTS', 'staging,production').split(',')
     end
   end
 end