+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# v1.8.0 (2017-07-24)
+# HEAD
 
 Features:
 
+- Adds pre-baked Google OAuth support (#44, #49)
 - adds 'default' to list of default Sidekiq queues
 - accepts custom Sidekiq queue names
 - accepts custom Sidekiq queue permitted latency values