+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 Features:
 
-- reports Sidekiq metrics only for queues defined in process
+- Adds pre-baked Google OAuth support (#44, #49)
+- reports Sidekiq metrics only for queues defined in process (#50)
 - adds 'default' to list of default Sidekiq queues
 - accepts custom Sidekiq queue names
 - accepts custom Sidekiq queue permitted latency values