+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-# v1.8.0 (2017-07-24)
+# HEAD
+
+Bug fixes:
+
+- Do not consider 'quiet' workers in the SLA sidekiq metric (#51)
 
 Features:
 
+- Adds pre-baked Google OAuth support (#44, #49)
+- reports Sidekiq metrics only for queues defined in process (#50)
 - adds 'default' to list of default Sidekiq queues
 - accepts custom Sidekiq queue names
 - accepts custom Sidekiq queue permitted latency values