+++ CHANGELOG.md
@@ -1,11 +1,22 @@
-# v1.8.0 (2017-07-24)
+# v1.8.0 (2017-07-26)
+
+Bug fixes:
+
+- Do not consider 'quiet' workers in the SLA sidekiq metric (#51)
 
 Features:
 
-- adds 'default' to list of default Sidekiq queues
-- accepts custom Sidekiq queue names
-- accepts custom Sidekiq queue permitted latency values
-- allows environment-specific application checks
+- Provides a `PLAYBOOK.md` template when detected missing (#42)
+- Adds pre-baked Google OAuth support (#44, #49)
+- Reports Sidekiq metrics only for queues defined in process (#50)
+- Finer-grained Sidekiq configuration (#46)
+    - adds 'default' to list of default Sidekiq queues
+    - accepts custom Sidekiq queue names and permitted latency values
+    - allows environment-specific application checks
+
+Other:
+
+- Fixes the test harness (#48)
 
 # v1.7.0 (2017-07-18)